requirements.txt
openai-whisper>=20230314
Eel>=0.14.0
wxPython>=4.2.1